split lsusb dumps as file getting out of hand
